LayoutTests/platform/playstation-ps5/media/media-source/media-mp4-hevc-bframes-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
Append media segment.
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
EXPECTED (sourceBuffer.buffered.length == '1'), OBSERVED '2' FAIL
EXPECTED (sourceBuffer.buffered.end(0) == '236'), OBSERVED '232.4' FAIL
END OF TEST
